requirements.txt
discord.py>=1.7.3
aiohttp<3.8.0,>=3.6.0
attrs>=17.3.0
chardet<5.0,>=2.0
async-timeout<4.0,>=3.0
yarl<2.0,>=1.0
typing-extensions>=3.6.5
multidict<7.0,>=4.5
idna>=2.0
PyYaml>=5.4.1